% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissC3Charts.R
\name{FrissC3ChartsOutput}
\alias{FrissC3ChartsOutput}
\title{Widget output function for use in Shiny}
\usage{
FrissC3ChartsOutput(outputId, chart, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{outputId of the component}

\item{chart}{charttype: either: FrissC3Gauge /}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Widget output function for use in Shiny
}

