% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissC3Charts.R
\name{FrissC3Gauge}
\alias{FrissC3Gauge}
\title{Friss C3 gauge}
\usage{
FrissC3Gauge(value = 50, min = 0, max = 100, text = "value",
  gaugeWidth = 20, color = "red", showMinMax = TRUE, width = NULL,
  height = NULL, transition = 250)
}
\arguments{
\item{value}{value to set the gauge on}

\item{min}{min value for the gauge}

\item{max}{max value for the gauge}

\item{text}{text to put under the gauge}

\item{gaugeWidth}{widh of the gauge within the div}

\item{color}{fill color of the gauge}

\item{showMinMax}{indicates wheater the min and max values are shown}

\item{width}{width of the div the gauge is put in}

\item{height}{height of the div the gauge is put in}
}
\description{
Friss C3 gauge
}

