% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissC3Charts.R
\name{FrissC3LineBarChart}
\alias{FrissC3LineBarChart}
\title{Friss Gauge}
\usage{
FrissC3LineBarChart(timeData, counts, lineData, color, showSubChart = FALSE,
  timeFormat = "\%Y-\%m-\%d", debounce = 250, width = "100\%",
  height = "400px", transition = 250, yLabel1 = "y-label",
  yLabel2 = "y-label2", xLabel = "x-label")
}
\description{
Friss custom c3.js HTMLWidget output binding
}

