% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissC3Charts.R
\name{FrissC3LineChart}
\alias{FrissC3LineChart}
\title{Friss C3 LineChart}
\usage{
FrissC3LineChart(timeData, lineData, color, showSubChart = FALSE,
  timeFormat = "\%Y-\%m-\%d", debounce = 250, width = "100\%",
  height = "400px", transition = 250, yLabel = "y-label",
  xLabel = "x-label")
}
\description{
Friss custom c3.js HTMLWidget output binding
}

