% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissC3Charts.R
\name{FrissC3LineChartOutput}
\alias{FrissC3LineChartOutput}
\title{C3 chart output function for use in Shiny}
\usage{
FrissC3LineChartOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{outputId of the component}

\item{width}{width of the chart}

\item{height}{height of the chart}

\item{chart}{charttype: either: FrissC3Gauge /}
}
\description{
C3 chart output function for use in Shiny
}

