% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissC3Charts.R
\name{FrissC3StackedAreaChart}
\alias{FrissC3StackedAreaChart}
\title{Friss Gauge}
\usage{
FrissC3StackedAreaChart(timeData, lineData, color, timeFormat = "\%Y-\%m-\%d",
  debounce = 250, width = "100\%", height = "400px", transition = 250)
}
\description{
Friss custom c3.js HTMLWidget output binding
}

