% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissC3Charts.R
\name{FrissC3StackedBarChart}
\alias{FrissC3StackedBarChart}
\title{Friss C3 LineChart}
\usage{
FrissC3StackedBarChart(barData, width = "100\%", height = "400px")
}
\description{
Friss custom c3.js HTMLWidget output binding
}

