% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissC3Charts.R
\name{renderFrissC3BarChart}
\alias{renderFrissC3BarChart}
\title{C3 Bar chart render function for use in Shiny}
\usage{
renderFrissC3BarChart(expr, env = parent.frame(), quoted = FALSE)
}
\description{
C3 Bar chart render function for use in Shiny
}

