% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissC3Charts.R
\name{renderFrissC3Charts}
\alias{renderFrissC3Charts}
\title{Widget render function for use in Shiny}
\usage{
renderFrissC3Charts(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

