% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissC3Charts.R
\name{renderFrissC3LineChart}
\alias{renderFrissC3LineChart}
\title{C3 linebarchart render function for use in Shiny}
\usage{
renderFrissC3LineChart(expr, env = parent.frame(), quoted = FALSE)
}
\description{
C3 linebarchart render function for use in Shiny
}

