% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FrissC3Charts.R
\name{renderFrissC3PieChart}
\alias{renderFrissC3PieChart}
\title{C3 gauge Widget render function for use in Shiny}
\usage{
renderFrissC3PieChart(expr, env = parent.frame(), quoted = FALSE)
}
\description{
C3 gauge Widget render function for use in Shiny
}

